fix a bug introduced when clean up the code.
git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codesv3@4613 eb38b3b0-33ea-0310-94ad-d43fea0d214f

+++ ON/ON-NEGF-share/allocate_states.c
@@ -35,8 +35,8 @@
 
 void allocate_states()
 {
-    int size;
-    size = ct.num_states * ct.num_kpts * (ct.spin_flag +1);
+    size_t size;
+    size = (size_t)ct.num_states * (size_t)ct.num_kpts * (size_t)(ct.spin_flag +1);
     //dprintf("\n dsdsd %d %d %d", ct.num_states, ct.num_kpts, ct.spin_flag);
     my_malloc(states, size, STATE);
     my_malloc(states1, size,  STATE);
